getting the first version of the python package running and including the first pytests
