+++ .pre-commit-config.yaml
@@ -1,0 +1,35 @@
+repos:
+-   repo: https://github.com/ambv/black
+    rev: 22.3.0
+    hooks:
+    - id: black
+      language_version: python
+-   repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v3.4.0
+    hooks:
+    -   id: check-yaml
+    -   id: trailing-whitespace
+    -   id: end-of-file-fixer
+        additional_dependencies: ['darglint', 'pydocstyle<4.0']
+-   repo: https://gitlab.com/pycqa/flake8
+    rev: 3.8.4  # pick a git hash / tag to point to
+    hooks:
+    -   id: flake8
+-   repo: https://github.com/asottile/reorder_python_imports
+    rev: v2.4.0
+    hooks:
+    -   id: reorder-python-imports
+-   repo: https://github.com/pre-commit/mirrors-mypy
+    rev: v0.971
+    hooks:
+    -   id: mypy
+        args: [--no-strict-optional, --ignore-missing-imports]
+-   repo: local
+    hooks:
+     -  id: tests
+        name: pytest
+        entry: pytest
+        pass_filenames: false
+        language: system
+        types: [python]
+        stages: [commit]
