--- local_csv.csv
@@ -1,11 +1,0 @@
-ID,NAME
-1,John
-2,Mary
-3,Peter
-4,Paul
-5,George
-6,Ringo
-7,John
-8,Paul
-9,George
-10,Ringo
