+++ requirements.txt
@@ -2,3 +2,4 @@ snowflake-connector-python
 snowflake-sqlalchemy
 python-dotenv
 pandas
+pytest
